Organize game utilities into subfolders
