+++ game/utils/levels/getGoalFromGridSize.ts
