+++ game/utils/levels/getLevelSettings.ts
